Add dumper script
This adds a dumper script that will download all the known GBFS files listed on GitHub. It is recommended that this script be run sparingly, as to avoid unnecessary downloading of files. Ideally, we would look at the TTL of each file we download and only download as necessary. However, that adds a lot of complexity to the script.

We have this installed and dumping data once every five minutes.