Better organization of data
The directory layout has been changed slightly, and now every run is being compressed to save disk space. The directory tree was changed so that each archive can be decompressed to the same directory and retain the same structure as when it went in, with the data flat in each date's folder. This makes the data much nicer to work with.