Streamline compilation
This commit introduces a new script into the compiler,
extract_and_compile.sh. This file means that it is now possible to skip
the extraction process before compiling, as it is now done on-the-fly.

For this to work, pipe a list of .tar.bz2 files into parallel, and run
this script for each of them.